BUG-#111: fix feed not updating
